requirements.txt
numpy
Pillow
torch
torchvision
tensorboard
tqdm
rich